+++ packages/sit-onyx/src/components/OnyxCheckbox/OnyxCheckbox.stories.ts
@@ -1,3 +1,4 @@
+import { createTruncationDecorator } from "@/utils/storybook";
 import { defineStorybookActionsAndVModels } from "@sit-onyx/storybook-utils";
 import type { Meta, StoryObj } from "@storybook/vue3";
 import OnyxCheckbox from "./OnyxCheckbox.vue";
@@ -64,5 +65,26 @@ export const WithoutLabel = {
   args: {
     ...Default.args,
     hideLabel: true,
+  },
+} satisfies Story;
+
+/**
+ * A checkbox with truncation. You can set the "truncation" property to choose between the different truncation types.
+ */
+export const WithTruncation = {
+  args: {
+    ...Default.args,
+    label: "Very long label that will be truncated",
+  },
+  decorators: [createTruncationDecorator("12rem")],
+} satisfies Story;
+
+/**
+ * This example shows a skeleton checkbox.
+ */
+export const Skeleton = {
+  args: {
+    ...Default.args,
+    skeleton: true,
   },
 } satisfies Story;