+++ packages/sit-onyx/src/components/OnyxRadioButton/OnyxRadioButton.vue
@@ -1,8 +1,14 @@
 <script lang="ts" setup generic="TValue">
 import { ref, watchEffect } from "vue";
+import OnyxSkeleton from "../OnyxSkeleton/OnyxSkeleton.vue";
 import type { RadioButtonProps } from "./types";
 
-const props = defineProps<RadioButtonProps<TValue>>();
+const props = withDefaults(defineProps<RadioButtonProps<TValue>>(), {
+  disabled: false,
+  required: false,
+  selected: false,
+  truncation: "ellipsis",
+});
 
 const selectorRef = ref<HTMLInputElement>();
 
@@ -10,8 +16,12 @@ watchEffect(() => selectorRef.value?.setCustomValidity(props.errorMessage ?? "")
 </script>
 
 <template>
-  <!-- TODO: decide on support prefix and/or folder -->
-  <label class="onyx-radio-button" :title="props.errorMessage">
+  <div v-if="props.skeleton" class="onyx-radio-button-skeleton">
+    <OnyxSkeleton class="onyx-radio-button-skeleton__input" />
+    <OnyxSkeleton class="onyx-radio-button-skeleton__label" />
+  </div>
+
+  <label v-else class="onyx-radio-button" :title="props.errorMessage">
     <!-- TODO: accessible error: https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Attributes/aria-errormessage -->
     <input
       ref="selectorRef"
@@ -23,11 +33,15 @@ watchEffect(() => selectorRef.value?.setCustomValidity(props.errorMessage ?? "")
       :checked="props.selected"
       :disabled="props.disabled"
     />
-    <span class="onyx-radio-button__label">{{ props.label }}</span>
+    <span class="onyx-radio-button__label" :class="[`onyx-truncation-${props.truncation}`]">
+      {{ props.label }}
+    </span>
   </label>
 </template>
 
 <style lang="scss">
+$input-size: var(--onyx-spacing-md);
+
 .onyx-radio-button {
   --onyx-radio-button-cursor: pointer;
   --onyx-radio-button-selector-border-color: var(--onyx-color-base-neutral-400);
@@ -38,8 +52,12 @@ watchEffect(() => selectorRef.value?.setCustomValidity(props.errorMessage ?? "")
 
   display: inline-flex;
   align-items: center;
-  height: 2.5rem;
+  max-width: 100%;
   cursor: var(--onyx-radio-button-cursor);
+
+  &:has(&__label) {
+    padding-right: var(--onyx-spacing-2xs);
+  }
 
   &:has(&__selector:hover) {
     --onyx-radio-button-selector-border-color: var(--onyx-color-base-primary-300);
@@ -103,7 +121,7 @@ watchEffect(() => selectorRef.value?.setCustomValidity(props.errorMessage ?? "")
       color: var(--onyx-radio-button-selector-outline-color);
       offset: 0;
     }
-    transition: outline 400ms;
+    transition: outline var(--onyx-duration-sm);
 
     border: {
       style: solid;
@@ -117,9 +135,9 @@ watchEffect(() => selectorRef.value?.setCustomValidity(props.errorMessage ?? "")
     display: inline-flex;
     justify-content: center;
     align-items: center;
-    width: var(--onyx-spacing-md);
-    min-width: var(--onyx-spacing-md);
-    max-width: var(--onyx-spacing-md);
+    width: $input-size;
+    min-width: $input-size;
+    max-width: $input-size;
     aspect-ratio: 1;
 
     &::before {
@@ -130,6 +148,24 @@ watchEffect(() => selectorRef.value?.setCustomValidity(props.errorMessage ?? "")
       background-color: var(--onyx-color-base-background-blank);
       border-radius: 100%;
     }
+  }
+}
+
+.onyx-radio-button-skeleton {
+  padding: var(--onyx-spacing-sm);
+  display: inline-flex;
+  align-items: center;
+  gap: var(--onyx-spacing-sm);
+
+  &__input {
+    height: $input-size;
+    width: $input-size;
+    border-radius: var(--onyx-radius-full);
+  }
+
+  &__label {
+    height: var(--onyx-spacing-md);
+    width: var(--onyx-spacing-3xl);
   }
 }
 </style>