+++ packages/sit-onyx/src/components/OnyxRadioButtonGroup/OnyxRadioButtonGroup.ct.tsx
@@ -1,12 +1,14 @@
 import { expect, test } from "../../playwright-axe";
 import type { SelectionOption } from "../OnyxRadioButton/types";
 import OnyxRadioButtonGroup from "./OnyxRadioButtonGroup.vue";
+import type { OnyxRadioButtonGroupProps } from "./types";
 
 const EXAMPLE_OPTIONS: SelectionOption<string>[] = [
   { label: "dummy.1", value: "1", id: "1" },
   { label: "dummy.2", value: "2", id: "2" },
   { label: "dummy.3", value: "3", id: "3" },
   { label: "dummy.4", value: "4", id: "4", disabled: true },
+  { label: "dummy.5", value: "5", id: "5", skeleton: true },
 ];
 
 test("should display correctly", async ({ mount, makeAxeBuilder, page }) => {
@@ -84,4 +86,27 @@ test("should display correctly when preselected", async ({ mount, makeAxeBuilder
 
   // ASSERT
   expect(accessibilityScanResults.violations).toEqual([]);
+});
+
+test("should truncate", async ({ mount }) => {
+  const options: OnyxRadioButtonGroupProps<string>["options"] = [
+    { label: "Very long label that will be truncated", id: "id-1" },
+    { label: "Very long required label that will be truncated", id: "id-2" },
+    {
+      label: "Very long label that will be truncated with multiline",
+      id: "id-3",
+      truncation: "multiline",
+    },
+  ];
+
+  // ARRANGE
+  const component = await mount(
+    <OnyxRadioButtonGroup
+      options={options}
+      headline="Truncated group headline"
+      style="max-width: 16rem;"
+    />,
+  );
+
+  await expect(component).toHaveScreenshot("truncation-vertical.png");
 });