+++ packages/sit-onyx/src/components/OnyxSwitch/OnyxSwitch.stories.ts
@@ -1,3 +1,4 @@
+import { createTruncationDecorator } from "@/utils/storybook";
 import { defineStorybookActionsAndVModels } from "@sit-onyx/storybook-utils";
 import type { Meta, StoryObj } from "@storybook/vue3";
 import OnyxSwitch from "./OnyxSwitch.vue";
@@ -43,5 +44,26 @@ export const Invalid = {
   args: {
     ...Default.args,
     errorMessage: "Error message",
+  },
+} satisfies Story;
+
+/**
+ * A switch with truncation. You can set the "truncation" property to choose between the different truncation types.
+ */
+export const WithTruncation = {
+  args: {
+    ...Default.args,
+    label: "Very long label that will be truncated",
+  },
+  decorators: [createTruncationDecorator("12rem")],
+} satisfies Story;
+
+/**
+ * This example shows a skeleton switch.
+ */
+export const Skeleton = {
+  args: {
+    ...Default.args,
+    skeleton: true,
   },
 } satisfies Story;