+++ packages/sit-onyx/src/components/OnyxSwitch/OnyxSwitch.vue
@@ -1,16 +1,19 @@
 <script lang="ts" setup>
 import { OnyxIcon } from "@/index";
-import checkSmall from "@sit-onyx/icons/check-small.svg?raw";
-import xSmall from "@sit-onyx/icons/x-small.svg?raw";
-import type { OnyxSwitchProps } from "./types";
 import { areObjectsFlatEqual } from "@/utils/comparator";
 import { transformValidityStateToObject } from "@/utils/forms";
+import checkSmall from "@sit-onyx/icons/check-small.svg?raw";
+import xSmall from "@sit-onyx/icons/x-small.svg?raw";
 import { computed, ref, toRefs, watch } from "vue";
+import OnyxSkeleton from "../OnyxSkeleton/OnyxSkeleton.vue";
+import type { OnyxSwitchProps } from "./types";
 
 const props = withDefaults(defineProps<OnyxSwitchProps>(), {
   modelValue: false,
   disabled: false,
   required: false,
+  truncation: "ellipsis",
+  skeleton: false,
 });
 
 const emit = defineEmits<{
@@ -55,7 +58,13 @@ watch(
 </script>
 
 <template>
+  <div v-if="props.skeleton" class="onyx-switch-skeleton">
+    <OnyxSkeleton class="onyx-switch-skeleton__input" />
+    <OnyxSkeleton v-if="!props.hideLabel" class="onyx-switch-skeleton__label" />
+  </div>
+
   <label
+    v-else
     class="onyx-switch"
     :class="{ 'onyx-required-marker': props.required, 'onyx-optional-marker': !props.required }"
   >
@@ -73,29 +82,44 @@ watch(
         <OnyxIcon :icon="isChecked ? checkSmall : xSmall" size="24px" />
       </span>
     </span>
-    <span v-if="!props.hideLabel" class="onyx-switch__label">{{ props.label }}</span>
+
+    <span
+      v-if="!props.hideLabel"
+      class="onyx-switch__label"
+      :class="[`onyx-truncation-${props.truncation}`]"
+    >
+      {{ props.label }}
+    </span>
   </label>
 </template>
 
 <style lang="scss">
+$container-padding: var(--onyx-1px-in-rem);
+$icon-size: 1.25rem;
+$input-width: calc(2 * $icon-size - 2 * $container-padding);
+
 .onyx-switch {
   display: inline-flex;
   align-items: center;
   cursor: pointer;
-  margin: var(--onyx-spacing-2xs);
-  height: 1.5rem;
+  gap: var(--onyx-spacing-2xs);
+  max-width: 100%;
 
   &__input {
     // position: absolute is needed here in order to hide the native checkbox.
     position: absolute;
     opacity: 0;
+    cursor: inherit;
+    width: 0;
+    height: 0;
+    margin: 0;
 
     &:checked + .onyx-switch__container {
       background-color: var(--onyx-color-base-primary-500);
 
       .onyx-switch__icon {
         background-color: var(--onyx-color-themed-neutral-100);
-        transform: translateX(0.875rem);
+        transform: translateX(calc(75% - $container-padding));
         color: var(--onyx-color-text-icons-primary-intense);
       }
     }
@@ -120,9 +144,7 @@ watch(
 
     &:invalid + .onyx-switch__container {
       background-color: var(--onyx-color-base-danger-200);
-      border: var(--onyx-1px-in-rem) solid var(--onyx-color-base-danger-500);
-      height: 1.375rem;
-      width: 2.25rem;
+      border-color: var(--onyx-color-base-danger-500);
 
       .onyx-switch__icon {
         background-color: var(--onyx-color-base-danger-500);
@@ -135,30 +157,28 @@ watch(
 
       .onyx-switch__icon {
         background-color: var(--onyx-color-base-background-blank);
-        transform: translateX(0.75rem);
         color: var(--onyx-color-text-icons-danger-intense);
       }
     }
   }
 
   &__container {
+    $width: calc(2 * $icon-size - 2 * $container-padding);
+
     display: inline-flex;
-    margin: var(--onyx-spacing-2xs);
-    width: 2.375rem;
-    height: 1.5rem;
+    width: $input-width;
+    min-width: $input-width;
+    padding: $container-padding;
+    box-sizing: border-box;
     background-color: var(--onyx-color-base-neutral-300);
     border-radius: var(--onyx-radius-full);
+    border: var(--onyx-1px-in-rem) solid transparent;
     transition: background-color var(--onyx-duration-sm) ease;
 
     .onyx-switch__icon {
       display: flex;
-      flex-wrap: wrap;
-      align-content: baseline;
       align-self: center;
-      justify-content: flex-end;
-      margin: var(--onyx-spacing-5xs);
-      width: 1.25rem;
-      height: 1.25rem;
+      justify-content: center;
       background-color: var(--onyx-color-themed-neutral-100);
       border-radius: var(--onyx-radius-full);
       transition:
@@ -166,14 +186,15 @@ watch(
         background-color var(--onyx-duration-sm) ease;
       overflow: hidden;
       color: var(--onyx-color-text-icons-neutral-soft);
+
+      .onyx-icon {
+        --icon-size: #{$icon-size};
+      }
     }
   }
 
   &__label {
-    padding: var(--onyx-spacing-2xs) 0;
-    overflow: hidden;
     color: var(--onyx-color-text-icons-neutral-intense);
-    text-overflow: ellipsis;
     font-family: var(--onyx-font-family);
     font-size: 1rem;
     font-style: normal;
@@ -225,6 +246,24 @@ watch(
     .onyx-switch__label {
       color: var(--onyx-color-text-icons-neutral-soft);
     }
+  }
+}
+
+.onyx-switch-skeleton {
+  display: inline-flex;
+  align-items: center;
+  gap: var(--onyx-spacing-2xs);
+
+  &__input {
+    // icon size + padding top/bottom + border top/bottom
+    height: calc($icon-size + 2 * $container-padding + 2 * var(--onyx-1px-in-rem));
+    border-radius: var(--onyx-radius-full);
+    width: $input-width;
+  }
+
+  &__label {
+    height: var(--onyx-spacing-md);
+    width: var(--onyx-spacing-3xl);
   }
 }
 </style>