+++ packages/sit-onyx/CHANGELOG.md
@@ -1,5 +1,41 @@
 # sit-onyx
 
+## 1.0.0-alpha.32
+
+### Minor Changes
+
+- c147e3d: feat: add property `skeleton` to OnyxButton, OnyxCheckbox, OnyxRadioButton and OnyxSwitch
+
+## 1.0.0-alpha.31
+
+### Minor Changes
+
+- 69f1569: feat: add truncation for OnyxButton, OnyxCheckbox, OnyxRadioButton and OnyxSwitch
+
+## 1.0.0-alpha.30
+
+### Patch Changes
+
+- c785597: fix: add padding-right for checkbox and radio button label
+
+## 1.0.0-alpha.29
+
+### Patch Changes
+
+- f04d083: fix(OnyxSwitch): update styles
+
+## 1.0.0-alpha.28
+
+### Minor Changes
+
+- 62c8d65: feat: add component `OnyxSkeleton`
+
+## 1.0.0-alpha.27
+
+### Minor Changes
+
+- 827a893: feat: create components `OnyxAppLayout` and `OnyxPageLayout`
+
 ## 1.0.0-alpha.26
 
 ### Minor Changes