+++ packages/vitepress-theme/CHANGELOG.md
@@ -1,5 +1,47 @@
 # @sit-onyx/vitepress-theme
 
+## 1.0.0-alpha.31
+
+### Patch Changes
+
+- Updated dependencies [c147e3d]
+  - sit-onyx@1.0.0-alpha.32
+
+## 1.0.0-alpha.30
+
+### Patch Changes
+
+- Updated dependencies [69f1569]
+  - sit-onyx@1.0.0-alpha.31
+
+## 1.0.0-alpha.29
+
+### Patch Changes
+
+- Updated dependencies [c785597]
+  - sit-onyx@1.0.0-alpha.30
+
+## 1.0.0-alpha.28
+
+### Patch Changes
+
+- Updated dependencies [f04d083]
+  - sit-onyx@1.0.0-alpha.29
+
+## 1.0.0-alpha.27
+
+### Patch Changes
+
+- Updated dependencies [62c8d65]
+  - sit-onyx@1.0.0-alpha.28
+
+## 1.0.0-alpha.26
+
+### Patch Changes
+
+- Updated dependencies [827a893]
+  - sit-onyx@1.0.0-alpha.27
+
 ## 1.0.0-alpha.25
 
 ### Patch Changes